indexer runs a little without raising exceptions
i've been drafting a structure for storing an index of the chain data, on the chain data, in such a way that it can be found without needing it, so as to reduce load on gateways.
the changes here are pretty messy, but it's a big accomplishment for me to get this far.